refactoring chip component
