+++ src/index.ts
@@ -4,7 +4,7 @@
 // https://material.angularjs.org/latest/demo/button
 // Core
 import Button, { ButtonProps } from './buttons/Button';
-import Chip, { ChipProps } from './chip/Chip';
+import Chip, { ChipProps } from './buttons/Chip';
 import Input, { InputProps } from './input/Input';
 import Icon, { IconProps } from './icons/Icon';
 import ListItem, { ListItemProps } from './list/ListItem';
